# pkg/bgpv1/gobgp: Optimize reconcileDiff String()