Avoiding to use  fmt.Sprintf() so that Go won't over-allocate the memory.

issue reference : cilium#19571

Result:

```
$ go test -benchmem -run=^$ -bench ^BenchmarkReconcileDiffString$ github.com/cilium/cilium/pkg/bgpv1/gobgp > BenchmarkReconcileDiffString_old.txt
$ go test -benchmem -run=^$ -bench ^BenchmarkReconcileDiffString$ github.com/cilium/cilium/pkg/bgpv1/gobgp > BenchmarkReconcileDiffString_new.txt
$ benchcmp  BenchmarkReconcileDiffString_old.txt BenchmarkReconcileDiffString_new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                           old ns/op     new ns/op     delta
BenchmarkReconcileDiffString-12     1592          350           -78.01%

benchmark                           old allocs     new allocs     delta
BenchmarkReconcileDiffString-12     18             5              -72.22%

benchmark                           old bytes     new bytes     delta
BenchmarkReconcileDiffString-12     264           152           -42.42%
```

Signed-off-by: MikeLing <sabergeass@gmail.com>