@ldelossa 2022-08-09
@MikeLing 

Are these optimizations worth wild? My assumption is that the String() method isn't called particularly often,  but measuring this would seem to be the first step to determine if this makes a difference or not?