@MikeLing 2022-08-09
>Are these optimizations worth wild? My assumption is that the String() method isn't called particularly often, but measuring this would seem to be the first step to determine if this makes a difference or not?

@ldelossa Measuring the called frequency make sense to me. Actually, like https://github.com/cilium/cilium/issues/19571#issuecomment-1203188316 mentioned, we. still have more than 200 places of code need to be changed from `fmt.Sprintf` to the others. Do we have any way to measure the function called frequency? Because some functions' called frequency are depend on the scenery. Like `String()`, we may need to call it every time if  the user need to inspect the cluster status, otherwise it won't be called like forever.