@christarazi 2022-08-30
@MikeLing Not all `fmt.Sprintf` instances need to be optimized as some of them are not in the hot path. I'd imagine this BGP code path is not necessarily in the hot path, but @ldelossa can confirm that.