@ldelossa 2022-08-30
Yes, I don't think we need to remove all instances of `fmt.Sprintf` and re-invent string building and the "%+v" format specifier. A better way to go about this would be to run Cilium in some common configurations and pprof `fmt.Sprintf` usages, then remove where they spin hot.